feat: add convert controller and service
Added the "yaml" package to convert JSON to YAML, since convert-js only seems to support YAML specs.
If we don't want to support a JSON spec on the HTTP request, we can delete this dependency.

If we do want that capability, then perhaps we should choose just one YAML package (yaml or js-yaml) that can do these conversions.

#13